+++ src/managednetworkfabric/azext_managednetworkfabric/tests/latest/test_l3domain.py
@@ -25,6 +25,7 @@ def call_scenario1(test):
     setup_scenario1(test)
     step_create(test, checks=[])
     step_show(test, checks=[])
+    step_update(test, checks=[])
     step_list_resource_group(test, checks=[])
     step_list_subscription(test, checks=[])
     step_update_admin_state_Enable(test, checks=[])
@@ -46,6 +47,12 @@ def step_show(test, checks=None):
         checks = []
     test.cmd(
         'az networkfabric l3domain show --resource-name {name} --resource-group {rg}')
+
+def step_update(test, checks=None):
+    '''l3domain update operation'''
+    if checks is None:
+        checks = []
+    test.cmd('az networkfabric l3domain update --resource-group {rg} --resource-name {name} --aggregate-route-configuration {updatedAggregateRouteConf}', checks=checks)
 
 def step_list_resource_group(test, checks=None):
     '''l3domain list by resource group operation'''
@@ -92,6 +99,7 @@ class GA_L3DomainScenarioTest1(ScenarioTest):
             'redistributeStaticRoutes': CONFIG.get('L3_ISOLATION_DOMAIN', 'redistribute_static_routes'),
             'connectedSubnetRoutePolicy': CONFIG.get('L3_ISOLATION_DOMAIN', 'connected_subnet_route_policy'),
             'aggregateRouteConf': CONFIG.get('L3_ISOLATION_DOMAIN', 'aggregate_route_conf'),
+            'updatedAggregateRouteConf': CONFIG.get('L3_ISOLATION_DOMAIN', 'updated_aggregate_route_conf'),
             'post_name': CONFIG.get('L3_ISOLATION_DOMAIN', 'post_name'),
             'state_Enable': CONFIG.get('L3_ISOLATION_DOMAIN', 'state_Enable'),
             'state_Disable': CONFIG.get('L3_ISOLATION_DOMAIN', 'state_Disable')